# Conflicts:
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/ws1617/iosl/pubcrawl20/Utilites/SignInHelper.java
@@ -272,7 +272,7 @@ public class SignInHelper {
 
   private void getCrawler(final GoogleSignInAccount acc) {
     SharedPreferences sharedPref = activity.getSharedPreferences(activity.getString(R.string.preference_user), Context.MODE_PRIVATE);
-    int id = sharedPref.getInt(activity.getString(R.string.user_id), -1);
+    long id = sharedPref.getLong(activity.getString(R.string.user_id), -1);
     if(id == -1) {
       setCrawlerID(acc);
     }
@@ -284,7 +284,7 @@ public class SignInHelper {
    * @param acc
    * @param id
    */
-  private void checkUser(GoogleSignInAccount acc, int id) {
+  private void checkUser(GoogleSignInAccount acc, long id) {
     Person person = null;
     try {
       person = new PersonDbHelper().getPerson(id);
@@ -306,7 +306,7 @@ public class SignInHelper {
       JSONObject selfJSON = linksJSON.getJSONObject("self");
       String link = selfJSON.getString("href");
       Log.i(TAG, link);
-      sharedPref.edit().putInt(activity.getString(R.string.user_id), getIDfromURL(link)).apply();
+      sharedPref.edit().putLong(activity.getString(R.string.user_id), getIDfromURL(link)).apply();
     } catch (JSONException e) {
       e.printStackTrace();
       return false;
